Update to latest cmd_system
Update BP32 dependencies for esp32
