+++ src/components/bluepad32/CMakeLists.txt
@@ -144,7 +144,7 @@ endif()
 
 # Valid for ESP-IDF only
 if(IDF_TARGET)
-    set(requires "nvs_flash" "btstack" "app_update" "esp_timer")
+    set(requires "nvs_flash" "btstack" "app_update" "esp_timer" "esp_driver_ledc" "esp_driver_spi" "driver")
 
     if("${IDF_VERSION_MAJOR}" GREATER_EQUAL 5)
         # When using ESP-IDF v5+, use "normal" components
